requirements.txt
beautifulsoup4==4.6.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
idna==2.7
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.10
lxml==4.2.4
MarkupSafe==1.0
newspaper3k==0.2.6
nltk==3.3
Pillow==5.2.0
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
requests-file==1.4.3
six==1.11.0
tldextract==2.2.0
urllib3==1.23
Werkzeug==0.14.1
